Fixed indentation and moved SchedulingConfiguration to a separate file
